mergeData.py ignores cleaning if multiple run

+++ analysis/scripts/mergeData.py
@@ -73,19 +73,19 @@ pwd = os.environ['PWD']
 
 bad_files = []
 
+ignored_cleaning_runs = ['ABC', 'ABCD']
+
 if i < 550:
   filename_out = outputDir+"/PhotonJet_2ndLevel_DATA_RUN_"+run+"_"+str(today)+".root" #+name[0]+"_"+name[1]+"_"+name[2]+"_"+name[3]+"_"+name[4]+"_"+str(today)+".root" 
   print filename_out
   exitcode = os.system("hadd -f "+filename_out+"  "+files1 +" > tmp_py_merge_files_from_step2_{}.out".format(inputList))
-  while exitcode > 0 and cleaning:
+  while exitcode > 0 and cleaning and not (run in ignored_cleaning_runs):
     print 'Error while merging, removing a file...'
     bad_file = os.popen("cat tmp_py_merge_files_from_step2_{}.out".format(inputList)).read()[:-1].split('\n')[-1].split(':')[1][1:]
     print bad_file
     bad_files.append(bad_file)
     if bad_file in files1:
       files1 = files1.replace(bad_file,'')
-    else:
-      import pdb; pdb.set_trace()
     exitcode = os.system("hadd -f "+filename_out+"  "+files1 +" > tmp_py_merge_files_from_step2_{}.out".format(inputList))
 else:
   filename_out = outputDir+"/PhotonJet_2ndLevel_DATA_RUN_"+run+"_"+str(today)+".root"#+name[0]+"_"+name[1]+"_"+name[2]+"_"+name[3]+"_"+name[4]+"_"+str(today)+".root" 
@@ -95,33 +95,29 @@ else:
   print filename_out2
   print filename_out
   exitcode = os.system("hadd -f "+filename_out1+" "+files1 +" > tmp_py_merge_files_from_step2_{}.out".format(inputList))
-  while exitcode > 0 and cleaning:
+  while exitcode > 0 and cleaning and not (run in ignored_cleaning_runs):
     print 'Error while merging, removing a file...'
     bad_file = os.popen("cat tmp_py_merge_files_from_step2_{}.out".format(inputList)).read()[:-1].split('\n')[-1].split(':')[1][1:]
     print bad_file
     bad_files.append(bad_file)
     if bad_file in files1:
       files1 = files1.replace(bad_file,'')
-    else:
-      import pdb; pdb.set_trace()
     exitcode = os.system("hadd -f "+filename_out1+" "+files1 +" > tmp_py_merge_files_from_step2_{}.out".format(inputList))
   exitcode = os.system("hadd -f "+filename_out2+" "+files2 +" > tmp_py_merge_files_from_step2_{}.out".format(inputList))
-  while exitcode > 0 and cleaning:
+  while exitcode > 0 and cleaning and not (run in ignored_cleaning_runs):
     print 'Error while merging, removing a file...'
     bad_file = os.popen("cat tmp_py_merge_files_from_step2_{}.out".format(inputList)).read()[:-1].split('\n')[-1].split(':')[1][1:]
     print bad_file
     bad_files.append(bad_file)
     if bad_file in files2:
       files2 = files2.replace(bad_file,'')
-    else:
-      import pdb; pdb.set_trace()
     exitcode = os.system("hadd -f "+filename_out2+" "+files2 +" > tmp_py_merge_files_from_step2_{}.out".format(inputList))
   os.system("hadd -f "+filename_out+"  "+filename_out1+" "+filename_out2 )
 
 all_files = os.popen("cat "+args.inputList).read()[:-1].split('\n')
 good_files = [f for f in all_files if f not in bad_files]
 
-if cleaning:
+if cleaning and not (run in ignored_cleaning_runs):
   os.system("rm "+args.inputList)
   for files in good_files:
     os.system("echo "+files+" >> "+args.inputList)
